# perf_invalid_flag_test.py:Adding testcase for perf segfaults