segfaults with invalid flag is observed
for command for 'perf --version -test'

Signed-off-by: Naveen kumar T <naveet89@in.ibm.com>