@naveen538 2019-09-04
Job Log:

Before Adding
------------------
avocado run perf_invalid_flag_test.py
JOB ID     : 4d5eacd5d3eea2960090aa73e6334cee538d598e
JOB LOG    : /home/workspace/runAvocadoFVTTest/avocado-fvt-wrapper/results/job-2019-09-04T10.33-4d5eacd/job.log
 (1/1) perf_invalid_flag_test.py:PerfInvalid.test_perf_invalid_flag: PASS (0.90 s)
RESULTS    : PASS 1 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB TIME   : 1.18 s
JOB HTML   : /home/workspace/runAvocadoFVTTest/avocado-fvt-wrapper/results/job-2019-09-04T10.33-4d5eacd/results.html


After Adding
---------------
avocado run perf_invalid_flag_test.py
JOB ID     : 4490f65115168ffb9c9191fba364479206157909
JOB LOG    : /home/workspace/runAvocadoFVTTest/avocado-fvt-wrapper/results/job-2019-09-04T10.34-4490f65/job.log
 (1/1) perf_invalid_flag_test.py:PerfInvalid.test_perf_invalid_flag: FAIL: perf: failed to execute command perf --version -test (0.94 s)
RESULTS    : PASS 0 | ERROR 0 | FAIL 1 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB TIME   : 1.29 s
JOB HTML   : /home/workspace/runAvocadoFVTTest/avocado-fvt-wrapper/results/job-2019-09-04T10.34-4490f65/results.html