@naveen538 2019-09-09
@PraveenPenguin Yes it worked.
job log :- /root/avocado/job-results/job-2019-09-09T00.43-4b1393f/job.log
ltc-zz189-lp6: 9.40.194.46